requirements.txt
Django==2.0.3
python-decouple==3.1
flake8==3.4.1
py==1.4.34
pytest==3.2.1
pytest-cov==2.3.0
pytest-django==3.1.2
dj-database-url
psycopg2
djangorestframework==3.7.7
